+++ ext_beacon_util.py
@@ -7,7 +7,7 @@
 #
 
 from scapy.all import *
-
+from rtap_ext_util import Listify_Radiotap_Headers
 
 
 
@@ -30,9 +30,6 @@ def return_IE_by_tag(pkt, tagno, list_enabled=False):
     input("")
 
   return ret_l
-
-
-
 
 
 class rates_descriptor_t:
@@ -116,24 +113,6 @@ class rates_descriptor_t:
   ###Notes of complex IE rates / modulations:
 
 
-## 1999 - 802.11b  2.4GHz, DSSS, 22MHz, 11MBps 
-## 1999 - 802.11a 5GHz, OFDM, 20MHz, 54Mbps
-## 2003 - 802.11g 2.4GHzm OFDM, 20MHz, 54Mbps
-#### Tag 1 (Rates):        Covers required rate   info from  1999-2003 inclusive 
-#### Tag 3 (DS Param set): Covers required chanel info from  1999-2003 inclusive
-
-### if (Tag1.Rates only goes to 11MBps) : 1999 802.11b   :  2.4
-### if (Tag1.Rates contains 54MBps)     : 2003 802.11g   : 
-### if (Tag3.channel > 11 (i.e. 149,etc): 2003 802.11a   : +5GHz
-
-
-## 2009 - 802.11n 2.4/5Ghz, MIMO-OFDM, 20,40mhz, 600MBps : +40MHz channels 4x4 mimo 
-## 2013 - 802.11AC 5GHz, MIMO-OFDM, 20,40,80,160MHZ,     : +80,160 MHz channsels 8x8 mimo
-
-### if (Tag61 present)                  : 2009 802.11N ()
-### if (Tag192 present)                 : 2013 802.11AC 
-
-## ?Critical IE
 class modulation_descriptor_t:
 
   rates_info = rates_descriptor_t()
@@ -222,20 +201,16 @@ class modulation_descriptor_t:
     return ret
 
 
-
-
 class TargetCharacteristics:
   # These values excep num_ext_antennas/related are parsed/deduced from Beacon Info Elements (not the radiotap meta header)
   _initialized = False
   tags = {}
 
-  #0
   ssid_hidden = False
   ssid= None
-  
-  
+
   modulation_info = modulation_descriptor_t()
-  
+
   num_ext_antennas = 0
   ext_antenna_list = []
   _inital_beacon = None
@@ -247,25 +222,23 @@ class TargetCharacteristics:
 
   def summary(self):
     ret = ""
-
     ret += ("Chan:%02d" %  self.modulation_info.curr_channel)
-
     if (self.ssid_hidden):
       ret += "SSID: <HIDDEN>"
     else:
       ret += " SSID: %s" % (self.ssid)
-    
-    #ret += "\n     Rates:%s" % (self.rate_info)
 
-    ret += "\n    Modulation:%s" % (self.modulation_info)
+    #ret += "    Rates:%s" % (self.rate_info)
+    ret += "    Modulation:%s" % (self.modulation_info)
     print(ret)
+    return ret
     
 
   def init_main(self, pkt):
     P=pkt.getlayer(Dot11)
     print("#### TargetCharacteristiscs::Interpret_Beacon_IEs::start")
-  
-    ## ID=00, SSID.             If SSID.len=0, or SSID is curiously completely missing, assume 'hidden' BSSID. 
+
+    ## ID=00, SSID.             If SSID.len=0, or SSID is curiously completely missing, assume 'hidden' BSSID.
     tag_0_ssid=return_IE_by_tag(pkt, 0)
     if (tag_0_ssid == None or tag_0_ssid.len == 0 or (tag_0_ssid.len == 1 and tag_0_ssid.info =="")):
         self.hidden_ssid = True
@@ -274,26 +247,15 @@ class TargetCharacteristics:
       self.tags[0]=tag_0_ssid
       self.ssid=tag_0_ssid.info.decode()
 
-   
+   ## ID=1,3,50, <HT/vht>
     self.modulation_info.process_pkt(pkt)
-    
-    ## ID=161,92, ... many for 802.11AC/N/G/
-    #self.modulation_info.process(pkt, self.rate_info, self.curr_channel) # Both the current channel (IE3) and supported rates (IE1,50)  req'd input
-    #### Okay, So: Channel, SSID, Rates/ExtendedRates are done.
-    #### What should be next?
-    #### 802.11 5Ghz/2Ghz detection? What does the channel say for an 11a network?
 
-   
-   
+
     ## ID=11: "QBSS Load element" (sta_count, channel utilization, )
-    ## ID=23: TPC (TransmitSignal Strength report? ?) #QQQ  
+    ## ID=23: TPC (TransmitSignal Strength report? ?) #QQQ
     ## ID=33:  IEEE80211_IE_POWER_CAPAB        33
-   ## ID=45: HT Capabilites (802.11N D1.10). Complicated.
+    ## ID=45: HT Capabilites (802.11N D1.10). Complicated.
     ##      \ TxBeamForming, AntennaSelection, HTCapabilites(20Mhz only, 40MHz intolerant), MCS Set, SecondarychjannelOffset
-    ##
-    ## 221/50:6f:9a: (WiFi alliance)/Type 9: WiFi alliance P2P info? 
-
-
     ## ID=72: "20/40 MHz BSS CoExistence info"
 
   def init(self, pkt): # Targetcharacteristics
@@ -301,46 +263,19 @@ class TargetCharacteristics:
     print("####TargetCharecteristics::init")
     R=pkt.getlayer(RadioTap)
     r=RadioTap( raw(R)[:R.len] )
-    
+
     #### Iterate through info-elements, storing/caching relevant info
-    self.init_main(pkt) 
+    self.init_main(pkt)
     print("#### 2) OK. Beacon data parsed. Shown bewlow")
     self.summary()
-    sys.exit(0)   
+    ######################################
 
-
-    print("####-TODO: following line, parse (at least hte 'top' level RTap headre)")
     ARrrs= Listify_Radiotap_Headers(pkt)
-    top_rtap = ARrrs.pop() 
+    top_rtap = ARrrs.pop()
     self.num_extra_measurements = len(ARrrs)
     print ("Num extra measurement on target:%d" % (self.num_extra_measurements))
-    print("##Channel precuror input: ")
 
-    idx=0
-    for p in ARrrs:
-      print("Ext-%d: AntSignal:(%d)" % (idx,p.dBm_AntSignal))
-      idx+=1
-
-def GetFirstBeacon(pkt):
-  print("####GetFirstBeacon::Start")
-  #ret = Listify_Radiotap_Headers(pkt)
-  #print("#####GetFirstBeacon::ListifyResults")
- 
-  print("----- Analyzing beacon into target characteristics")
-  T = TargetCharacteristics()
-  T.init(pkt)
-  
-  input("###^^How does that look in terms of meta info?")
-  sys.exit(0)
-  #T.init(pkt)
-  #ssid=str(pkt.getlayer(Dot11).info) #XXX This conveniently contains SSID (IELement 0. But this isnt a great approach)
-  #print("SSID: %s" % (ssid))
-  #input("")
-  #return ssid
-  #print(ret)
-  #wrpcap("out2.pcap", ret)
-  #sys.exit(0)
-
+############# ext_beacon_util::main ####
 def main():
   ## Misc platform setup: On Macos we need to explicitly enable libpcap for BPF to work correctly
   conf.use_pcap = True
@@ -378,35 +313,7 @@ def main():
      sniff(prn=A.Simpl_Process_Radiotap, iface=A.Config.input_src, filter=bpfilter, monitor=1, store=0, count=0)
 
 
-#  sniff(prn=A.Simpl_Process_Radiotap, offline="png.pcap", filter=bpfilter, monitor=1, store=0, count=0)
-
 
 
 if __name__=='__main__':
   main()
-#tcpdump  -I -i png.pcap  type mgt subtype beacon
-## Here we put a BPF filter so only 802.11 Data/to-DS frames are captured
-#s = conf.L2listen(offline = IN_INPUT,filter = "link[0]&0xc == 8 and link[1]&0xf == 1")
-## TODO Here we _should_ put a BPF filter so only 802.11 Management frames are captured
-## XXX JC: RESUME HERE ^^^. Fix BSSID, investigate pkts /w no signal,
-# tcpdump  -I -i png.pcap  -e  type mgt subtype beacon  -c 10 -w ~/t.cap3
-
-
-# Target: The AWUS1900 / 88XXau.ko driver
-#         This device supports a/b/g/n/ac with 4 Antennas.
-#		  On top of this, the driver provides a radiotap antenna tag _per_ antenna
-#
-
-# # The snark is strong in this one: https://dox.ipxe.org/ieee80211_8h_source.html
-#  /** 802.11 Robust Security Network ("WPA") information element
-#   781  *
-#   782  * Showing once again a striking clarity of design, the IEEE folks put
-#   783  * dynamically-sized data in the middle of this structure. As such,
-#   784  * the below structure definition only works for IEs we create
-#   785  * ourselves, which always have one pairwise cipher and one AKM;
-#   786  * received IEs should be parsed piecemeal.
-#   787  *
-#   788  * Also inspired was IEEE's choice of 16-bit fields to count the
-#   789  * number of 4-byte elements in a structure with a maximum length of
-#   790  * 255 bytes.
-#   791  *
